Fix the rest of the higan/icarus bug
Removed the check for is_file from the consoles page. Why? Well...
	higan folders are not files, but can be opened!
	So we treat them as they are a rom!
Added a check for higan autogen files (files ending on .rom or .ram)
Works flawlessly with my system currently.

TODO: (maybe) Add a check that only does the higan check if the console
or emulator is detected as higan or a higan console.